backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Microsoft Integration
msal==1.26.0

# Monitoring & Logging
prometheus-client==0.19.0
psutil==5.9.7
python-json-logger==2.0.7

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0
